--- src/experiments/content-translation/types/window.ts
@@ -1,10 +1,0 @@
-/**
- * Internal dependencies
- */
-import type { AIContentTranslationData } from './types';
-
-declare global {
-	interface Window {
-		aiContentTranslationData: AIContentTranslationData;
-	}
-}
